feat: production-readiness remediation (audit, blockers, follow-ups, close-out)
Security:
- untracked secrets (.env, prefs, docker envs); env-first AI keys; fail-closed internal auth
- tracking-credential + settings-API encryption at rest; GPS truck-ownership checks
- tenant isolation: per-company GPS queues, cleanup/retention scoping, insight dedup

Reliability:
- GPS flush: retry + delete-after-commit; email double-send dedup; OCR status field
- export None-guards; desktop global excepthook; rotating logs; JSON backend logs
- freight rate-limiter/circuit-breaker wiring; tracking backoff/jitter; tacho version probe

Performance:
- dashboard/board WorkerPool; batched driver/route resolves; composite (company_id, start_date) index
- OTD metric (promised_date) end-to-end incl. desktop form; batched conflict scan + ISO date parsing

Quality:
- i18n: t() coverage, RO translations, 22-file key parity; dependency CVE pins + lockfiles
- coverage baseline + CI parallel-mode fix; ~35 pre-existing test failures triaged (2 real bugs: transaction commit batching, phantom cmr_audit index)
- docs: production-readiness audit + deploy runbook

Note: working tree contained extensive pre-existing uncommitted WIP; unrelated items (website/, env files, scratch scripts, report docs, untracked support/mobile modules) were excluded from this commit.

+++ alembic/versions/a7b8c9d0e1f7_create_copilot_insights.py
@@ -18,7 +18,7 @@ def upgrade() -> None:
     op.create_table(
         'copilot_insights',
         sa.Column('id', sa.UUID(), primary_key=True, server_default=sa.text('gen_random_uuid()')),
-        sa.Column('company_id', sa.Integer(), sa.ForeignKey('companies.id'), nullable=False),
+        sa.Column('company_id', sa.BigInteger(), sa.ForeignKey('companies.id'), nullable=False),
         sa.Column('insight_type', sa.Text(), nullable=False),
         sa.Column('payload', sa.JSON(), nullable=False),
         sa.Column('severity', sa.Text(), nullable=False, server_default='low'),
